class_02/pyproject.toml
[project]
name = "class-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
